Move update schedule page to /internal/updates with sidebar nav
- Move page from /wiki/updates to /internal/updates so it gets the
  internal layout with left sidebar navigation
- Add "Update Schedule" to Dashboards & Tools section in internal nav
- Update PageStatus frequency link to point to /internal/updates
- Match internal page style (prose max-w-none wrapper)
